% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass.R
\name{Bass}
\alias{Bass}
\title{Bass model}
\usage{
Bass(x, h = 0, optim = F, estim.met = c("nls", "ols"))
}
\arguments{
\item{x}{vector of non-cumulative observed data to fit Bass model on}

\item{h}{desired step-ahead forecast (to become depreciated) [default = 0]}

\item{optim}{additional optimisation on the insample MSE}

\item{estim.met}{non-linear least squares (\code{"nls"}) or ordinary least
squares (\code{"ols"}) [default "nls"]}
}
\value{
\code{coef}: coefficients p, q and m

list of fitted values:
\itemize{
\item \code{fit} fitted values
\item \code{actuals} actual values
\item \code{out} predicted values
\item \code{RMSE} insample root mean squared error
}
}
\description{
\code{Bass} fits a Bass model (Bass 1969)
}
\details{
The Coefficients are estimated by first fitting a linear regression 
  as proposed by Bass (1969). This then serves as starting values for the 
  non-linear least square estimation (Srinivasan and Mason 1986).
  Furthermore, parameters can be additionally optimised on the MSE of the
  insample following the suggestion of Lilien et al. (2000) using BOBYQA
  algorithm (Powell 2009). If p and q can not be determined optimal region,
  they are manually set to values of either 0.00000001 or 0.99999999 for p
  and 0.00000001 or 1.49999999 for q, depending whether the proposed values
  of the optim()-function are above or below of 0 to 1 or 0 to 1.5,
  respectively.
}
\references{
Bass, F.M., 1969. A new product growth for model consumer 
  durables. Management Science 15(5), 215-227.

Srinivasan, V., and Mason, C.H., 1986. Nonlinear Least Squares
  Estimation of New Product Diffusion MOdels. Marketing Science, 5(2),
  169-178.

Lilien, G.Y., Rangaswamy, A. and Van den Bulte, C., 2000. 
  Diffusion models: Managerial applications and software. In: Mahajan, V., 
  Muller, E., Wind, Y. (Eds.), New-Product Diffusion Models. Kluwer, 
  Massachusetts, 295-311.

Powell, M.J.D., 2009. The QOBYQA algorithm for bound constrained 
  optimization witouth derivatives. Department of Applied Mathematics and 
  Theorteical Physics, Cambridge University. DAMTP 2009/NA06.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
