% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_broadband.R
\docType{data}
\name{broadband}
\alias{broadband}
\title{Broadcast subscribers}
\format{A data frame with 51 observations and 4 variables
\describe{
\item{time}{Quarters}
\item{DSL}{Adoption of DSL subscribers}
\item{CableModem}{Adoption of CableModem users}
\item{FTTPb}{Adoption of FTTPb sales}
  }}
\usage{
broadband
}
\description{
A dataset containing the broadcast subscribers to UK market 4Q 2011 to 2Q 2016.
}
\references{
Telecoms Market Matrix
}
\keyword{datasets}
