% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chicken.R
\docType{data}
\name{chicken}
\alias{chicken}
\title{Chicken weight}
\format{A data frame with 13 observations and 2 variables
\describe{
\item{time}{Weeks since birth}
\item{weight}{Weight of the female chicken in Kg}
  }}
\usage{
chicken
}
\description{
A dataset containing the average weekly female chicken weight.
}
\references{
Jukic, D., Kralik, G. and Scitovski, R. (2004). Least-square
  fitting Gompertz curve. Journal of Computational and Applied Mathematics,
  169, 359-375.
}
\keyword{datasets}
