% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{Gompertz}
\alias{Gompertz}
\title{Gompertz model}
\usage{
Gompertz(x, startval = NULL)
}
\arguments{
\item{x}{vector of cumulative observations}

\item{startval}{an optional Vector with starting for manual estimation}
}
\value{
coefficients for a, b and c
}
\description{
\code{Gompertz} fits a Gompertz model using non-linear least square
estimation.
}
\details{
For \code{startval} the Vector values needs to be named in the form
  of (a, b, c). Else automated approximation of NLS is taking place using
  Jukic et al. (2004) approach.
}
\examples{
# Gompertz model example --------------------------------------------------
\dontrun{
data("chicken")
Gompertz(x = chicken$weight)

data("carstock")
Gompertz(carstock$raw)
}
}
\references{
Jukic, D., Kralik, G. and Scitovski, R., 2004. Least-squares
  fitting Gompertz curve. Journal of Computational and Applied Mathematics,
  169, 359-375.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
