% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{is.diffusion}
\alias{is.diffusion}
\alias{is.bass}
\title{Diffusion class checkers}
\usage{
is.diffusion(x)

is.bass(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if this is the specified class and \code{FALSE} otherwise.
}
\description{
Functions to check if an object is of the specified class
}
\details{
The list of functions includes:
\itemize{
\item \code{is.diffusion()} tests if the object was produced by a diffusion() function.
\item \code{is.bass()} checks if the forecast was produced by the bass() function.
}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru},

Oliver Schaer, \email{info@oliverschaer.ch}
}
\keyword{ts}
\keyword{univar}
