% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nortonbass.R
\name{Nortonbass_error}
\alias{Nortonbass_error}
\title{Fits Norton Bass curve and estimated RMSE}
\usage{
Nortonbass_error(x, param, gstart = NULL, flexpq = F)
}
\arguments{
\item{x}{matrix with generations}

\item{param}{the parameters for curve to estimated}

\item{gstart}{optional vector of starting points for the generations}

\item{flexpq}{flexible p and q}
}
\value{
yhat, the predicted values

actuals, the actual values

RMSE, the root mean squared error for each generation
}
\description{
Fits Norton Bass curve and estimated RMSE
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
\keyword{internal}
