% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdiffusion.R
\name{plot.seqdiffusion}
\alias{plot.seqdiffusion}
\title{Plot sequentially fitted diffusion curves.}
\usage{
\method{plot}{seqdiffusion}(x, cumulative = c(FALSE, TRUE), ...)
}
\arguments{
\item{x}{\code{seqdiffusion} object, produced using \code{\link{seqdiffusion}}.}

\item{cumulative}{If TRUE plot cumulative adoption.}

\item{...}{Unused argument.}
}
\value{
None. Function produces a plot.
}
\description{
Produces a plot of sequentially fitted diffusion curves.
}
\examples{
 fit <- seqdiffusion(tsIbm)
 plot(fit)

}
\seealso{
\code{\link{seqdiffusion}}.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch},

Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
\keyword{internal}
