% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{print.diffusion}
\alias{print.diffusion}
\title{Print a fitted diffusion curve.}
\usage{
\method{print}{diffusion}(x, ...)
}
\arguments{
\item{x}{\code{diffusion} object, produced using \code{\link{diffusion}}.}

\item{...}{Unused argument.}
}
\value{
None. Console output only.
}
\description{
Outputs the result of a fitted diffusion curve.
}
\examples{
 fit <- diffusion(tsChicken[, 2])
 print(fit)

}
\seealso{
\code{\link{diffusion}}.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch},

Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
