% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdiffusion.R
\name{print.seqdiffusion}
\alias{print.seqdiffusion}
\title{Print sequentially fitted diffusion curves.}
\usage{
\method{print}{seqdiffusion}(x, ...)
}
\arguments{
\item{x}{\code{seqdiffusion} object, produced using \code{\link{seqdiffusion}}.}

\item{...}{Unused argument.}
}
\value{
None. Console output only.
}
\description{
Outputs the result of sequentially fitted diffusion curves.
}
\examples{
 fit <- seqdiffusion(tsIbm)
 print(fit)

}
\seealso{
\code{\link{seqdiffusion}}.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch},

Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
\keyword{internal}
