% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdiffusion.R
\name{seqdiffusion}
\alias{seqdiffusion}
\title{seqdiffusion}
\usage{
seqdiffusion(x, cleanlead = c(TRUE, FALSE), prew = NULL, l = 2,
  pvalreps = 0, eliminate = c(FALSE, TRUE), sig = 0.05,
  verbose = c(FALSE, TRUE), type = c("bass", "gompertz", "sgompertz"))
}
\arguments{
\item{x}{matrix containing in each column the adoption per period for generation k}

\item{cleanlead}{removes leading zeros for fitting purposes (default == T)}

\item{prew}{the \code{w} of the previous generation. This is used for
sequential fitting.}

\item{l}{the l-norm (1 is absolute errors, 2 is squared errors)}

\item{pvalreps}{bootstrap repetitions to estimate (marginal) p-values}

\item{eliminate}{if TRUE eliminates insignificant parameters from the
estimation. Forces \code{pvalreps = 1000} if left to 0.}

\item{sig}{significance level used to eliminate parameters}

\item{verbose}{if TRUE console output is provided during estimation (default
== F)}

\item{type}{diffusion model to use}

\item{w}{vector of model parameters (see note). If provided no estimation
is done.}
}
\value{
list of:
\itemize{
\item \code{type} diffusion model type used
\item \code{diffusion} returns model specification for each generation (see
\code{\link{diffusion}} for details)
\item \code{call} calls function fitted
\item \code{w} named matrix of fitted parameters for each generation
\item \code{x} matrix of actuals
\item \code{mse} insample Mean Squared Error for each generation
\item \code{pval} all p-values for \code{w} at each generation
}
}
\description{
\code{seqdiffusion} enables to fit various sequential diffusion models.
}
\details{
This function fits diffusion models of the type \code{"bass"}, 
\code{"gompertz"} or \code{sgompertz} across generations. Parameters are 
estimated for each generation individually by minimising the Mean Squared 
Error with the subplex algorithm from the nloptr package. Optionally p-values
of the coefficients can be determined via bootstraping. Furthermore, the
bootstrapping allows to remove insignificant parameters from the optimisation
process.
}
\note{
vector \code{w} needs to be provided for the Bass model in the order of
  \code{"p", "q", "m"}, where "p" is the coefficient of innovation, "q" is the
  coeficient of imitation and "m" is the market size coefficient.
  
  For the Gompertz model vector \code{w} needs to be in the form of
  \code("a", "b", "m"). Where "a" is the x-axis displacement coefficient, "b"
  determines the growth rate and "m" sets, similarly to Bass model, the
  market potential (saturation point).
}
\examples{
# example for seqdiffusion() function
  data("ibm")
  fitsgomp <- seqdiffusion(ibm, type = "sgompertz")
  fitgomp <- seqdiffusion(ibm, type = "gompertz")
  fitbass <- seqdiffusion(ibm, type = "bass")
  
  plot(fitgomp)
  plot(fitsgomp)
  plot(fitbass)
  
}
\references{
Bass, F.M., 1969. A new product growth for model consumer
  durables. Management Science 15(5), 215-227.

Bemmaor, A. 1994. Modeling the Diffusion of New Durable Goods:
  Word-of-Mouth Effect versus Consumer Heterogeneity. In G. Laurent, G.L.
  Lilien and B. Pras (Eds.). Research Traditions in Marketing. Boston:
  Kluwer, pp. 201-223.

Jukic, D., Kralik, G. and Scitovski, R., 2004. Least-squares
  fitting Gompertz curve. Journal of Computational and Applied Mathematics,
  169, 359-375.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch},

Nikoloas Kourentzes, \email{nikoloas@kourentzes.com}
}
