% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsac}
\alias{tsac}
\title{Time series: Assassins Creeds}
\format{A matrix with 380 observations and 8 variables
\describe{
\item{ac1}{Assassins Creed 1}
\item{ac2}{Assassins Creed 2}
\item{ac3}{Assassins Creed 3}
\item{ac4}{Assassins Creed 4}
\item{ac5}{Assassins Creed 5}
\item{ac6}{Assassins Creed 6}
\item{ac7}{Assassins Creed 7}
\item{ac8}{Assassins Creed 8}
  }}
\description{
A dataset containing the weekly sales of Assassins Creeds game.
}
\references{
VGChartz
}
