% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsbroadband}
\alias{tsbroadband}
\title{Time series: Broadcast subscribers}
\format{A data frame with 51 observations and 4 variables
\describe{
\item{time}{Quarters}
\item{DSL}{Adoption of DSL subscribers}
\item{CableModem}{Adoption of CableModem users}
\item{FTTPb}{Adoption of FTTPb sales}
  }}
\usage{
tsbroadband
}
\description{
A dataset containing the broadcast subscribers to UK market 4Q 2011 to 2Q 2016.
}
\references{
Telecoms Market Matrix
}
\keyword{datasets}
