% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tscarstock}
\alias{tscarstock}
\title{Time series: Stock of cars}
\format{A data frame with 25 observations and 3 variables
\describe{
\item{year}{Year}
\item{raw}{Raw stock numbers}
\item{smoothed}{Smoothed stock numbers as described by Franses (1994)}
  }}
\description{
A dataset containing the yearly stock of cars in the Netherlands (1965-1989).
}
\references{
Franses, P.H. (1994). Fitting a Gompertz curve. Journal of
  Operational Research Society, 45, 109-113.
}
