% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsCovid}
\alias{tsCovid}
\title{Time series: COVID-19 confirmed cases US}
\format{
A ts object with 107 days of observations
\describe{
\item{tsCovid}{Daily confirmed COVDID-19 cases}
  }
}
\source{
\url{https://github.com/CSSEGISandData/COVID-19}
}
\description{
A dataset containing the number of confirmed COVID-19 cases in the US.
}
\references{
COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University
}
