% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsIbm}
\alias{tsIbm}
\title{Time series: Sales of IBM Computers}
\format{
A data frame with 24 observations and 4 variables \describe{
\item{SIU1}{1st generation}
\item{SIU2}{2nd generation (starts 6 years after first generation)}
\item{SIU3}{3rd generation (starts 11 years after first generation)}
\item{SIU4}{4th generation (starts 16 years after first generation)}}
}
\source{
\url{https://citeseerx.ist.psu.edu/document?repid=rep1&type=pdf&doi=8bbf197bc39a27ccf44cfd5ed22b5db3da0c7bb2}
}
\description{
A dataset containing the first four generations of yearly IBM general-purpose
computers installations in the USA.
}
\references{
Bass, P.I. and Bass, F.M. 2004. IT Waves: Two Completed
  Generational Diffusion Models. Working Paper Basseconomics, 1-33.
}
