% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsMetal}
\alias{tsMetal}
\title{Time series: U.S. Merchant Marine conversion to metal}
\format{A data frame with 17 observations and 2 variables
\describe{
\item{year}{Year}
\item{substitution}{Conversion to metal}
  }}
\description{
A dataset with conversion of U.S. Merchant Marine from wood to metal.
}
\references{
Martino, J.P. 1993. Technological Forecasting for Decision
  Making. 3rd edition. New York: McGraw-Hill.
}
