% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass-Model.R
\name{bass}
\alias{bass}
\title{Statistical Bass Model}
\usage{
bass(
  data,
  lags = frequency(data),
  seasonality = FALSE,
  distribution = c("dlnorm", "dgamma", "dinvgauss"),
  formula = NULL,
  ...
)
}
\arguments{
\item{data}{A vector or a matrix with variables. In the latter case if
\code{formula} is not provided then it is assumed that the response variable
is in the first column of the data.}

\item{lags}{The vector of seasonal lags. Only required if \code{seasonality=TRUE}.}

\item{seasonality}{The logical variable. If \code{TRUE} then the seasonal model is
estimated. The seasonality is modelled in the style of ETS model.}

\item{distribution}{The type of distribution to assume for the error term. Can be
\code{"dlnorm"} - Log-Normal, \code{"dgamma"} - Gamma or \code{"dinvgauss"} -
Inverse Gaussian. "d" in the name of distribution specifies that we use density
function.}

\item{formula}{R formula specifying which variable to include in the model and
in what form.}

\item{...}{Parameters passed to the optimiser (nloptr):
\itemize{
\item \code{maxeval} - maximum number of evaluations to carry out. The default is 40 per
estimated parameter for ETS and / or ARIMA and at least 1000 if explanatory variables
are introduced in the model (100 per parameter for explanatory variables, but not less
than 1000);
\item \code{maxtime} - stop, when the optimisation time (in seconds) exceeds this;
\item \code{xtol_rel} - the relative precision of the optimiser (the default is 1E-6);
\item \code{xtol_abs} - the absolute precision of the optimiser (the default is 1E-8);
\item \code{ftol_rel} - the stopping criterion in case of the relative change in the loss
function (the default is 1E-8);
\item \code{ftol_abs} - the stopping criterion in case of the absolute change in the loss
function (the default is 0 - not used);
\item \code{algorithm} - the algorithm to use in optimisation
(by default, \code{"NLOPT_LN_SBPLX"} is used);
\item \code{print_level} - the level of output for the optimiser (0 by default).
If equal to 41, then the detailed results of the optimisation are returned.
}
You can read more about these parameters by running the function
\link[nloptr]{nloptr.print.options}.
Starting values of parameters can be passed via \code{B}, while the upper and lower
bounds should be passed in \code{ub} and \code{lb} respectively. In this case they
will be used for optimisation.}
}
\value{
Object of class "bass" is returned. It contains the list of the
following values:

\itemize{
\item \code{data} - the provided data.
\item \code{timeElapsed} - time elapsed for the construction of the model.
\item \code{fitted} - the fitted values, shifted in time.
\item \code{residuals} - the residuals of the model.
\item \code{B} - the vector of estimated parameters.
\item \code{curves} - adoption, innovators and imitators curves.
\item \code{scale} - the value of the estimated scale parameter.
\item \code{distribution} - the assumed distribution.
\item \code{loss} - type of loss function used in the estimation.
\item \code{lossValue} - loss function value.
\item \code{logLik} - log-likelihood of the model.
}
}
\description{
Function estimates the statistical Bass model
}
\details{
The function accepts the data and applies the Bass model with
assumed distribution and defined components to it. This is based on
the paper by Svetunkov et al. (202??). The model assumes that the market
size variable m is stochastic and might change over time.
}
\examples{

# CMA of specific order
ourModel <- bass(tsWindows[,6])

}
\references{
\itemize{
\item{Bass, F.M., 1969. A new product growth for model consumer durables. Management Science 15(5), 215-227.}
}
}
\seealso{
\code{\link[smooth]{es}, \link[smooth]{ssarima}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{ts}
\keyword{univar}
