% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{diffusion}
\alias{diffusion}
\title{diffusion}
\usage{
diffusion(x, w = NULL, cleanlead = c(TRUE, FALSE), prew = NULL, l = 2,
  pvalreps = 0, eliminate = c(FALSE, TRUE), sig = 0.05,
  verbose = c(FALSE, TRUE), type = c("bass", "gompertz", "sgompertz"))
}
\arguments{
\item{x}{vector with adoption per period}

\item{w}{vector of model parameters (see note). If provided no estimation
is done.}

\item{cleanlead}{removes leading zeros for fitting purposes (default == T)}

\item{prew}{the \code{w} of the previous generation. This is used for
sequential fitting.}

\item{l}{the l-norm (1 is absolute errors, 2 is squared errors)}

\item{pvalreps}{bootstrap repetitions to estimate (marginal) p-values}

\item{eliminate}{if TRUE eliminates insignificant parameters from the
estimation. Forces \code{pvalreps = 1000} if left to 0.}

\item{sig}{significance level used to eliminate parameters}

\item{verbose}{if TRUE console output is provided during estimation (default
== F)}

\item{type}{diffusion model to use}
}
\value{
returns list of:
\itemize{
\item \code{type} diffusion model type used
\item \code{call} calls function fitted
\item \code{w} named vector of fitted parameters
\item \code{x} actuals
\item \code{fit} fitted values of model
\item \code{mse} insample Mean Squared Error
\item \code{prew} the \code{w} of the previous generation
\item \code{pval} p-values for \code{w}
}
}
\description{
\code{diffusion} enables to fit various diffusion models.
}
\details{
This function fits diffusion models of the type \code{"bass"}, 
\code{"gompertz"} or \code{"sgompertz"}. Parameters are estimated by
minimising the Mean Squared Error with a quasi Newton algorithm. Optionally
p-values of the coefficients can be determined via bootstraping. Furthermore,
the bootstrapping allows to remove insignificant parameters from the
optimisation process.
}
\note{
vector \code{w} needs to be provided for the Bass model in the order of
  \code{"p", "q", "m"}, where "p" is the coefficient of innovation, "q" is the
  coeficient of imitation and "m" is the market size coefficient.
  
  For the Gompertz model vector \code{w} needs to be in the form of
  \code("a", "b", "m"). Where "a" is the x-axis displacement coefficient, "b"
  determines the growth rate and "m" sets, similarly to Bass model, the
  market potential (saturation point).
}
\examples{
# example for diffusion() function
\dontrun{
  
  # Some examples needed
  
}
}
\references{
Bass, F.M., 1969. A new product growth for model consumer
  durables. Management Science 15(5), 215-227.

Jukic, D., Kralik, G. and Scitovski, R., 2004. Least-squares
  fitting Gompertz curve. Journal of Computational and Applied Mathematics,
  169, 359-375.
}
\seealso{
\code{\link{seqdiffusion}} for sequential diffusion model fitting
  across product generations.
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch},

Nikoloas Kourentzes, \email{nikoloas@kourentzes.com}
}
