% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nortonbass.R
\name{Nortonbass_startvalgen}
\alias{Nortonbass_startvalgen}
\title{Fits Norton Bass curve and estimated RMSE}
\usage{
Nortonbass_startvalgen(x, gstart, flexpq, startval.met)
}
\arguments{
\item{x}{matrix with generations}

\item{gstart}{optional vector of starting points for the generations}

\item{flexpq}{For \code{"startvalgen="BB"}. Allows parameters p and q to be flexible if set \code{TRUE}.}

\item{startvalgen}{\code{"iBM"} fits individual Bass model to each generation;
\code{"BB"} uses the approach described in Bass and Bass (2004).}
}
\value{
starting values for all parameters
}
\description{
Fits Norton Bass curve and estimated RMSE
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
\keyword{internal}
