% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tsibm}
\alias{tsibm}
\title{Time series: Sales of IBM Computers}
\format{A data frame with 24 observations and 4 variables \describe{
\item{SIU1}{1st generation}
\item{SIu2}{2nd generation (starts 6 years after first generation)}
\item{SIu3}{3rd generation (starts 11 years after first generation)}
\item{SIu4}{4th generation (starts 16 years after first generation)}}}
\source{
\url{http://www.bassbasement.org/F/N/BBDL/Bass and Bass 2004 AS.pdf}
}
\usage{
tsibm
}
\description{
A dataset containing the first four generations of yearly IBM general-purpose
computers installations in the USA.
}
\references{
Bass, P.I. and Bass, F.M., 2004. IT Waves: Two Completed
  Generational Diffusion Models. Working Paper Basseconomics, 1-33.
}
\keyword{datasets}
